refactor: add VC embed functions into views dir
