+++ src/events/buttons/manageVotecount/players/addPlayers.ts
@@ -1,10 +1,9 @@
 import { type UserSelectMenuBuilder, type ButtonBuilder, type ButtonInteraction, type CacheType, ActionRowBuilder } from 'discord.js';
 import { CustomButton } from '../../../../structures/interactions/Button';
 import { getVoteCounter } from '@models/votecounter';
-import { generateManagePlayersEmbed } from '../gotoPlayersMenu';
 import AddPlayersMenu from '../../../selectMenus/manageVotecount/addPlayers';
 import { UserSelectMenu } from '../../../../structures/interactions/UserSelectMenu';
-import { manageVoteCountEmbeds } from '../goHome';
+import { genCreateVoteCountEmbed, genPlayersEmbed } from '@views/votecounter';
 
 export default class AddPlayersButton extends CustomButton {
 	static customId = 'manage-vc-players-add';
@@ -14,9 +13,8 @@ export default class AddPlayersButton extends CustomButton {
 
 	async onExecute(i: ButtonInteraction<CacheType>) {
 		const vc = await getVoteCounter({ channelId: i.channelId });
-		if (!vc) return manageVoteCountEmbeds.create();
-		const payload = generateManagePlayersEmbed(vc);
-
+		if (!vc) return genCreateVoteCountEmbed();
+		const payload = genPlayersEmbed(vc);
 		const row = new ActionRowBuilder<UserSelectMenuBuilder>();
 
 		const select = UserSelectMenu.getUserSelectMenuOrThrow(AddPlayersMenu.customId);
