+++ src/events/buttons/manageVotecount/state/jumpToDay.ts
@@ -1,9 +1,9 @@
 import { type ButtonBuilder, type ButtonInteraction, type CacheType } from 'discord.js';
 import { CustomButton } from '../../../../structures/interactions/Button';
 import { getVoteCounter } from '@models/votecounter';
-import { generatePlaceholder } from '../gotoPlayersMenu';
 import { Modal } from '../../../../structures/interactions/Modal';
 import SetDayModal from '../../../modals/setDay';
+import { genCreateVoteCountEmbed } from '@views/votecounter';
 
 export default class JumpToDayButton extends CustomButton {
 	static customId = 'manage-vc-players-jump-to-day';
@@ -13,7 +13,7 @@ export default class JumpToDayButton extends CustomButton {
 
 	async onExecute(i: ButtonInteraction<CacheType>) {
 		const vc = await getVoteCounter({ channelId: i.channelId });
-		if (!vc) return i.update(generatePlaceholder(vc ?? undefined));
+		if (!vc) return i.update(genCreateVoteCountEmbed());
 
 		// Unhandled error
 		const setDayModal = Modal.getModalOrThrow(SetDayModal.customId);
