+++ src/events/selectMenus/manageVotecount/removePlayers.ts
@@ -1,9 +1,8 @@
 import { type UserSelectMenuInteraction, type CacheType } from 'discord.js';
 import { UserSelectMenu } from '../../../structures/interactions/UserSelectMenu';
 import { InteractionError } from '../../../structures/interactions';
-import { generateManagePlayersEmbed } from '../../buttons/manageVotecount/gotoPlayersMenu';
-import { manageVoteCountEmbeds } from '../../buttons/manageVotecount/goHome';
 import { deletePlayerAndVotes, getVoteCounter } from '@models/votecounter';
+import { genCreateVoteCountEmbed, genPlayersEmbed } from '@views/votecounter';
 
 export default class RemovePlayersMenu extends UserSelectMenu {
 	static customId = 'manage-vc-players-remove';
@@ -18,7 +17,7 @@ export default class RemovePlayersMenu extends UserSelectMenu {
 		if (!values) throw new InteractionError('No users were provided');
 
 		const vc = await getVoteCounter({ channelId: i.channelId });
-		if (!vc) return i.update(manageVoteCountEmbeds.create());
+		if (!vc) return i.update(genCreateVoteCountEmbed());
 
 		const players = vc.players;
 
@@ -50,8 +49,8 @@ export default class RemovePlayersMenu extends UserSelectMenu {
 		}
 
 		const newVC = await getVoteCounter({ channelId: i.channelId });
-		if (!newVC) return i.update(manageVoteCountEmbeds.create());
-		const playerMenuPayload = generateManagePlayersEmbed(newVC);
+		if (!newVC) return i.update(genCreateVoteCountEmbed());
+		const playerMenuPayload = genPlayersEmbed(newVC);
 		await i.update(playerMenuPayload);
 	}
